requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.31.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
uuid-utils==0.14.1
